+++ __init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-"""zero2gpkg_converter plugin factory.
+"""zero2cadgis plugin factory.
 100% English.
 """
 
 
 def classFactory(iface):
-    from .main_plugin import Zero2GpkgConverter
-    return Zero2GpkgConverter(iface)
+    from .main_plugin import Zero2CadGis
+    return Zero2CadGis(iface)